Refactor notifications code
