+++ re_data/notifications/slack.py
@@ -1,4 +1,7 @@
+from tabulate import tabulate
 import requests
+from typing import Any, Dict, Optional, List, Tuple
+from datetime import datetime
 
 
 def slack_notify(webhook_url: str, body: dict) -> None:
@@ -16,3 +19,126 @@ def slack_notify(webhook_url: str, body: dict) -> None:
 
     response = requests.post(webhook_url, json=body, headers=headers)
     response.raise_for_status()
+
+def format_alerts_to_table(alerts: list, limit=None) -> str:
+    """
+    Formats a list of alerts to a table.
+    :param alerts:
+        List of alerts exported from dbt-re-data.
+    :return: str
+        Formatted table.
+    """
+    table = []
+    for alert in alerts:
+        table.append([
+            # alert['model'],
+            alert['message'],
+            alert['value'],
+            alert['time_window_end'],
+        ])
+    if limit: 
+        table = table[:limit]
+    return tabulate(table, headers=['Message', 'Value', 'Time Window'], tablefmt='simple')
+
+
+def generate_slack_message(model, details, owners, subtitle: str, selected_alert_types: set) -> dict:
+    """
+    Generates a slack message for a given model.
+    """
+    anomalies = details['anomalies']
+    schema_changes = details['schema_changes']
+    tests = details['tests']
+    slack_owners = [k[0] for k in owners]
+    message_obj = {
+        'blocks': [
+            {
+                "type": "header",
+                "text": {
+                    "type": "plain_text",
+                    "text": "Model: {}".format(model),
+                    "emoji": True
+			    }
+            },
+            {
+                "type": "divider"
+            },
+            {
+                "type": "section",
+                "text": {
+                    "type": "mrkdwn",
+                    "text": "Owners: {}".format(', '.join(slack_owners))
+                }
+            },
+            {
+                "type": "divider"
+            },
+            {
+                "type": "section",
+                "fields": [
+                    {
+                        "type": "plain_text",
+                        "text": ":warning: {} anomalies".format(len(anomalies)),
+                        "emoji": True
+                    },
+                    {
+                        "type": "plain_text",
+                        "text": ":bulb: {} schema changes".format(len(schema_changes)),
+                        "emoji": True
+                    },
+                    {
+                        "type": "plain_text",
+                        "text": ":bangbang: {} failed tests".format(len(tests)),
+                        "emoji": True
+                    }
+                ]
+            },
+        ]
+    }
+    if anomalies and 'anomaly' in selected_alert_types:
+        message_obj['blocks'].append({
+            "type": "section",
+            "text": {
+                "type": "mrkdwn",
+                "text": "*Anomalies*\n ```{}```".format(format_alerts_to_table(anomalies, limit=10))
+            }
+        })
+    if schema_changes and 'schema_change' in selected_alert_types:
+        message_obj['blocks'].append({
+            "type": "section",
+            "text": {
+                "type": "mrkdwn",
+                "text": "*Schema Changes*\n ```{}```".format(format_alerts_to_table(schema_changes, limit=10))
+            }
+        })
+    if tests and 'test' in selected_alert_types:
+        message_obj['blocks'].append({
+            "type": "section",
+            "text": {
+                "type": "mrkdwn",
+                "text": "*Tests failures*\n ```{}```".format(format_alerts_to_table(tests, limit=10))
+            }
+        })
+    if subtitle:
+        message_obj['blocks'].append({
+			"type": "section",
+			"text": {
+                "type": "mrkdwn",
+                "text": subtitle
+            }
+		}
+    )
+    message_obj['blocks'].append({
+			"type": "context",
+			"elements": [
+				{
+					"type": "plain_text",
+					"text": "Generated at {}. Check re_data UI for more details".format(datetime.now().strftime('%Y-%m-%d %H:%M:%S')),
+					"emoji": True
+				}
+			]
+		}
+    )
+
+    return message_obj
+    
+
